use traces.New to create trace instead of manually doing that in the otel converter
